requirements.txt
asgiref==3.9.1
certifi==2025.10.5
charset-normalizer==3.4.4
cloudinary==1.44.1
dj-database-url==3.0.1
Django==5.2.6
django-cloudinary-storage==0.3.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==11.3.0
psycopg==3.2.12
psycopg2-binary==2.9.11
PyJWT==2.10.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
django-cors-headers
cloudinary
django-cloudinary-storage